[UNI-131] feat: Route id 기반 탐색 구현

+++ uniro_backend/src/main/java/com/softeer5/uniro_backend/route/repository/RouteRepository.java
@@ -12,7 +12,7 @@ import org.springframework.stereotype.Repository;
 import com.softeer5.uniro_backend.route.entity.Route;
 
 @Repository
-public interface RouteRepository extends JpaRepository<Route, Integer> {
+public interface RouteRepository extends JpaRepository<Route, Long> {
 
     @EntityGraph(attributePaths = {"node1", "node2"})
     @Query("SELECT r FROM Route r WHERE r.univId = :univId")
